pyproject.toml
[project]
name = "ctrl-alt-bot"
version = "0.1.0"
description = "Mom's Friendly Robots Company approved bots."
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "flashbots>=2.0.0",
    "python-dotenv>=1.1.1",
    "questionary>=2.1.1",
    "web3==6.20.2",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]
